requirements.txt
# Core:
Django==1.10

# REST
djangorestframework==3.3.3
django-rest-swagger==0.3.8
